chore: extract deepEqual to separate file
