+++ packages/vuetify/src/util/index.ts
@@ -4,6 +4,7 @@ export * from './bindProps'
 export * from './colorUtils'
 export * from './console'
 export * from './createSimpleFunctional'
+export * from './deepEqual'
 export * from './defineComponent'
 export * from './dom'
 export * from './easing'
